model/requirements.txt
azure-ai-documentintelligence
azure-storage-blob
python-dotenv
